Added a method for finishing a task
The finishing task method is called separately in the method of each task to allow for things to happen after the task has completed but before the next task has started.

For example, we want to log exactly when a knot is detected and mark this as the end of the task, but we do not want to switch to the next task straight away.

There will most likely be a better way to handle this in the future, but it should work for now.