Imported VR client from Extraction repo
These files have been copied from the rrw/handling-shared-state branch of the Extraction repo. Development will now continue in this repo. This will facilitate connecting the puppeteering client and the VR client.

Current behaviour of the game (if a local server is running):
1. The player starts in a black space with the molecular simulation and a grey sphere visible.
2. The player makes an index finger pinch to centre the position of the sphere in front of their headset.
3. The play puts their hand in the sphere, and it starts to shrink.
4. Once the sphere has disappeared a game menu will appear, centred at the position of the sphere.
5. The user can click through the UI using the buttons, some of which will write information to the shared state.
If there is no local server running, no simulation will be visible and there will be no shared state to write to, but the behaviour is otherwise unchanged.

The README file has been updated with instructions about which Unity version to install and how to run the game.