Create new puppeteering client that waits for a VR client to connect
The puppeteering client is a class that connects to a local Nanover server and can write the shared state.

The puppeteer writes to the shared state that the game is ready to start and then waits for a VR client to connect ('player.Connected' key in the shared state).  Once connected, the puppeteer writes that the game is in progress to the shared state, waits 1 second, and ends the game by writing to the shared state and disconnecting from the server.

The key-value pairs that are written to the shared state have the namespace "puppeteer" to distinguish these keys from the ones that will be written by the VR client.