Starting game button only starts game when the hands are currently tracking, otherwise a warning appears
On the game intro menu, the "let's go" button will only move onto the knot tying task menu if clicked while the hands are tracking. If it is clicked when the controllers are tracking, a warning message will appear, along with a button to watch a video on how to enable hand tracking in the Oculus settings. When the user clicks this button (with either hands or controllers), they will be taken to a menu with the video with buttons to (a) go back, (b) play/replay the button, and (c) exit the game (so that they can go change the settings in their headset).

This is an alternative to automatically detecting if hand tracking is enabled, which is not currently possible for a game that allows both hands and controllers.