Turn on passthrough when sphere is active, turn off when it's not
Passthrough with edge rendering is enabled when the sphere gameobject is active. Passthrough turns off when the sphere is disabled (at which point the menu UI appears and the user starts/carries on with the game).

Passthrough must be enabled in the Oculus PC app to use it when running the game in Unity. Instructions for doing this have been added to the README.