Passthrough now enabled for Game Intro
This allows the player to put down there controllers if they are using them, since hand tracking needs to be enabled to continue with the game.

The passthrough manager handles the logic such that (a) if none of the gameobjects with the PassthroughToggle.cs script attached are active then passthrough will be disabled, and (b) if any of the gameobjects are active then passthrough should be enabled.