Sphere intro is now a menu canvas and can be switched using CanvasSwitcher.cs
This means that we no longer need the canvas closer, because now that the Sphere Intro has a "CanvasType", all the required logic can now be handled using the CanvasSwitcher script.

Having multiple menus that required passthrough broke the previous logic, so the passthrough logic has been updated to fix these issues.

Issues still to fix: the background of the canvas for the UI is too light when passthrough is enabled. The sphere material also needs to be updated.

The GameManager was integrated into the PuppeteerManager and is not needed anymore.