Improved the UI
Moved the menu canvas further from the headset, but kept the button closer to the player so that it is still within reach. The first button to press now says "press me!" to make it clear to the player that it is a button to be pressed. This then sets the precedent for the rest of the buttons.

The sphere now has a toggle that means that the sphere will only shrink if the hands are placed in it (if the handsOnly bool is toggled on) or if the controllers are placed in it (if the handsOnly bool is toggled off).