Removed handler scripts that were used to write to the shared state and add listener to puppeteer manager for updates in the shared state
The share state handlers were used to write to the shared state, but these scripts are not being used at the moment.

The Puppeteer Manager now listens for updates in the shared state dictionary. This cannot be done directly via the simulation.Multiplayer.SharedStateDictionary since this dictionary does not get populated.

The CanvasTextModifier modifies the text for the sphere intro depending on what the current interaction modality is in the shared state.