Removed 'end' as a separate task and added ids to CanvasType enum
Previously, 'end' was defined as a task in the shared state by the puppeteering client. Now the VR client gets all of the tasks defined by in the shared state and adds the Game End at the end of the list.

I added ids to the CanvasType enum so that I can add new values in the middle of the list without resetting the selected CanvasType in each menu canvas in Unity.