Pressing the game intro menu hides the menu and shows the simulation, allowing the player to interact with it
When the player clicks the "press me" button on the game intro menu, the client connects to the server, hides the menu, shows the simulation and activates the players interaction with it.

In the NarupaImdScript, for some reason the ConnectionEstablished?.Invoke(); line does not seem to be invoking the connected functions, so for the moment there is a workaround whereby a bool "serverConnected" is set for the simulation, and the PinchGrab script looks at that value instead.