Added the Enabling Hand Tracking menu
This has the same functionality as before, where the button to access this menu appears when the 'press me!' button is pressed with controllers. The player can then click the new button to move to a menu, where they can play a video that demonstrates how to enable hand tracking in the Oculus settings on the headset.

The name of the function to replay the video was updated to be more specific. The video player now pauses when the game starts, rather than stopping. This means that there is a preview of the video when the menu appears.