Update the way that the canvases are handled
The UI for each task now has a single canvas, which is sectioned into different menus. The canvas manager is responsible for changing between canvases and menus.

The button controller now does not know which canvas needs to be next, this logic is all handled by the canvas manager.

Everything that needs to happen at the start of the game is now handled by the appropriate manager, rather than nested into random code...