Updated the logic of the switching between canvases and menus
The each button can now either switch to the next menu (within a single task), or switch canvas (when the player is switching task).  The order of the menus within a task is stored in the CanvasController attached to the canvas of each set of menus.

In the puppeteering client, the nanotube task has been added as a separate class, which inherits from the 'Task' type.