Show/hide simulation logic fixed and the player must wait until the vr client has received the list of tasks from the server
The show/hide simulation logic was previously disabling the puppeteering client, not the simulation object. Now it hides the simulation object.

There is a second Intro Menu, which acts as a buffer to give time to the VR client to connect to the server. Once the puppeteer manager has an order of tasks, the start button will appear on the menu and the user can start the first task (currently this is only the nanotube task).

The ButtonController script has been cleaned up a bit, with consistent documentation and consistent logic for checking if the buttons can only be pressed when hand tracking is enabled.