The vr client now waits for the grabbers to be initialised as well as the server to be connected and  hardcoded IP address
The "User Interaction" object must be active when the server connected in order for the pinch grabbers to be initialised. The Puppeteer Manager now turns on user interactions to initialise the pinch grabbers (via the PinchGrab scrip), after which the user interactions are disabled again until required. This fixes the bug before which meant that the player could not interact with the simulations (because the pinch grabbers were never initialised).

The button to start the first task now only appears once the server is connected AND the pinch grabbers have been initialised.

Also, an IP is hardcoded into the Puppeteer Manager so that the VR client connects to the correct server. The ConnectionEstablished?.Invoke() does not seem to work anymore when using simulation.Connect() rather than Autoconnect(). The script now uses a bool to verify if the server has been connected. This is not as neat but works for now.