Added knot-tying task to the puppeteering client
This implements Knot Pull, which is a program made by dzarmola and is available on GitHub (https://github.com/dzarmola/knot_pull). When the task is running, it will check for a knot in the simulation approximately 30 times per second.

Note that KnotPull will only work on polypeptides and would need to be adapted for use with other types molecules.