Refactored the way that the simulation box is centred in front of the player
This was previously being done by the PuppeteerManager, but is now handled by a separate script (SimulationBoxCentre.cs). The PuppeteerManager calls the function at the same point as previously.

The script puts the simulation in front of the player by using the centre of the box and applying an offset relative to the forward direction of the headset (using the Center Eye Anchor).