Add loading of specific simulation
The puppeteering client now reads the list of simulations that have been loaded onto the server. The index corresponding to the correct simulation is given to the relevant task to be loaded onto the server.

There is a slight delay in receiving the frames when the task is first started, so a check is made in the Task class to make sure that the frames are being received. From testing, only one cycle of this loop is required before the frames are received.