Initial trial task
The puppeteering client has a timer for the task (10 seconds), after which it pauses the simulation and logs this to the shared state.  This is registered by the VR client, which automatically submits a dummy answer.

There are still issues, e.g. with the UI, but this is a first pass.