Update buckyball simulation
Buckyball simulations are now centred inside a box with sides of length 5nm. The molecules have been rotated such that they are aligned along the x-axis, which is the same as the nanotube sim. This means that both simulations are in the desired position in front of the player when the simulation box is automatically placed in front of them.