Log if answer is correct or not
The simulation files written by the customisable_openmm_system.py script have the following format: "buckyball_{type of force constant modified}_{residue ID of modified molecule}_{force constant multiplier}.xml". The puppeteer client now determines what the correct answer is for the current trial based on the name of the simulation file.

The 2 buckyball simulation file has been renamed to follow this format, although it remains to be the simulation with two identical buckyballs (no force constants have been modified, as  is specified by the multiplier of "1" written in the file name).

The VR client writes the clients submitted answer to the shared state, and the puppeteering client reads this, wipes the key from the shared state, and writes the correctness of this answer to the shared state. The correctness is either "True" (correct answer), "False" (incorrect answer) or "None" (the molecules are identical so there is no correct answer).