Change input sim index to sim indices
Previously, the simulation index that was being passed into the task was an integer (since only one simulation was needed per task). Now we need to pass in multiple indices for the trials, so the simulation indices variable is now a list, and the default index that is used for the each task is 0 (this sorts out the nanotube + knot-tying tasks), but we can pass different indices for the trials task.