Loop through number of alpha carbons, not residues
There was an issue where sometimes the number of residues was not the same as the number of alpha carbons. This should not be the case, but it is more rigorous to run through the number of alpha carbons anyway. Knot detection still works well from testing it in VR.