Tell the player which interaction modality they will be using
This is told to the player on a menu screen at the beginning of the nanotube task (as this will always be at the beginning of the two sections).

The player can now press the button to start the nanotube task with either modality, since they have not yet been told which one they will be using. From then onwards, they can only press buttons and interact with the molecule using the correct modality.