Write 'ready' to shared state only when the frames are being received
There was an issue where the previous simulation would flash up when loading a new task. The VR client now waits for the 'puppeteer.task_status' key to be 'ready' before showing the simulation. The pupeteering client only sets this key once the simulation is playing and the frames are being received.