Compile lists of strings used in the puppeteering client
This list is at the top of the puppeteering_client.py script and will be used to standardise the strings that are used throughout these python scripts.

In this commit we also standardised the server name and the interaction modality have already been changed in the initialiser of the puppeteering client.