Write avatar data to the shared state
The VR client now writes the positions and rotations of the centre eye anchor, right hand anchor and left hand anchor to the shared state 30 times a second. These are currently formatted in a very basic way as a proof of concept, but this can be changed easily when we know what we want to do with the data.