First pass over trials task + refactoring loading of simulations
At the beginning of the Trials task, it sorts through the loaded buckyball simulations and chooses one simulation for each multiplier value (randomly choosing between simulations in which A is altered and B is altered).

This required some refactoring of the way that simulations are loaded on the nanotube and knot-tying tasks.

The knot-tying task was also using the first_frame to get the resids and particle names, but this refers to the first frame of the first simulation loaded on the server, not the first frame of the currently loaded simulation. This wasn't an issue until now because it just happened that 17-Ala was always being loaded first on the server because it is first in the list in the directory.

This is very much a first pass and not the final code, the code needs to be cleaned up a lot, particularly the Trials task class.