Update player task status when changing task
Previously the Player.TaskStatus key in the shared state was only ever set to "InProgress" and "Finished". This meant that there was no task status for the intro to the first task and then each subsequent intro the task status remained "Finished". This has not been fixed.

Set default TaskStatusValue to None otherwise the first time it is set to Intro it does not get written to the shared state.