Show in-task instructions on back of the simulation box
An instructions panel now shows anytime the simulation is visible. This is placed on the back of the simulation box (on the XY plane) relative to the player.

The instructions show either how to interact with the molecules using controllers or hands, depending on which interaction modality is currently being used.