Add icons for scoring trials
There is a trial icon prefab, which has four child gameobjects. These four game objects are images representing the icon in the following states: normal, correct, incorrect and ambivalent. These images are enabled depending on whether the player gives the correct answer in the trials.

There are seven icons because there will be seven trials per set in the final game.

Features that need to be added:
- Disable the icons when not in the trials task
- Start from icon 1 again when doing the trials for the second time and disable the images representing the previous answers