Show final trial icon of each set
The icons were being reset at the beginning of the trials task and after each set. This meant that the icon for the final trial in each set was reset before the player had time to see their answer.

Now the icons are reset at the beginning of the task and, for all but the first set, they are reset after the answer submission for the first trial.