Only update position of box when switching task
The position of the simulation box now updates only (a) when the simulation box values update and (b) when the player is at the beginning of a new task. Previously, the simulation box position was moving at the beginning of each new trial, which was (a) annoying and disconcerting and (b) was problematic if the player moved a lot during a trial, since the UI could end up outside of the play space.