Check for stable knot
The code now checks that a knot has been formed over 30 consecutive frames. This seems to work well, but the number 30 was chosen arbitrarily, so maybe this should change in the future.

In terms of the UI, nothing has changed for the moment. Luis suggested that this is probably best, but we can see if this is strange. So the user is not notified when the first knot is detected, the simulation only changes colour once the stable knot is detected, and then the confetti comes a little after that.