Added timer for nanotube and knot-tying tasks
Added a timer for the nanotube and knot-tying tasks. Currently the timer just counts down from the beginning of the task and then stays at zero. Still need to add the logic for finishing the task once the timer is up. The limit is set to 10 seconds for both tasks at the moment (for testing only).