Swap exit sandbox to button on UI
This is to make it easier for players to exit the sandbox. The menu button on the controller is not easy for novice players to press, nor is the left hand finger pinch. Providing this button also allows players to get used to the presence of a button here, which will be good training for the trials task.