Create new version of interaction mode menu
This the first draft of the interaction mode menu using the new UI + on a single menu (previously 2 menus).

All of the information that was on the previous version of this menu is on the new menu, but hasn't been arranged on the canvas, so the text + images are overlapping.