Update interaction mode icon on progress bar
Update the interaction mode icon based on the current interaction mode set in the game manager. Note that we added "None" to the enums for the task type and input mode, because the default value for enums is the first value in the list of enums (not None or null), which means that the icons are not always set properly. The default is now "None" so that when the task and interaction mode are first set, they are updated.