Icons now align in grid and chips placed above menus
A grid layout group has been added to the ProgressChipHandler and the parameters have been adjusted to be able to fit 12 (current/future task) icons nicely. Note that the current task icons are bigger than the completed task icon. The position is set manually in code to be just above the wide menu UI.