All icons are initialised at the beginning of the game
The entire list of tasks is initialised at the beginning of the game, so the player can see all of the tasks they are required to do. However, the logic for updating a particular icon is broken because of the new grid layout group, meaning that updated chips get put to the end of the set of icons.