Wipe trials answer from shared state + fix code for progress chips
For some reason the trials answers were being left in the shared state and picked up at the beginning of the next trials task. Now the VR client wipes the answer from the shared state.

The ProgressChipHandler.cs script has been fixed - the counting of tasks wasn't previously taking into account the fact that training trials + trials are separate tasks for the game manager, but not for the chips.