Updated trials videos
Updated the trials videos to use the Quest 3 controllers and to show good technique. The sizes of the videos are slightly different, so this was changed in the scene.

Next to do: add videos for mudra gloves (already made) and implement logic to show the video corresponding to the current interaction mode.