Initial working trials observer task
The tasks runs and the player can see the recording. The shared state is not streaming (the avatar and interactions are not visible, but this may be a recording issue). Note that both of the recordings are identical, but one has been labelled 1.5 for testing (they are both the 0.5 multiplier).