Added switching interaction mode icon + redadded interaction mode icons
A switching interaction mode icon is now placed in the center of the progress bar. The icon is the interaction mode that the player will switch to at that point in the game.

The icons for the hands + controllers for the current task progress chip had been renamed + were missing in the prefab. These have been re-added.