Add short delay to starting sandbox
This delay (0.75 seconds) gives time for the server to switch simulation. Otherwise, the player sees the previous simulation (frozen) for a split second before the alkanes load.

Also, load the sandbox simulation first so that if the player does see the simulation for a split second, it will be the correct one.