Add a script for putting a game object at the position where we want the headset to be relative to the box
This script will be attached to a game object at the same level as the simulation box. It sets the position of the game object to be at a specific offset (depending on the current task) relative to the box center. The position of this object will then be used to calibrate the space.