Replace old script for positioning sim box with a CalibratedSpaceController, and hook up objects in scene
Replaced the old PositioningSimBox script with a CalibratedSpaceController script. This script calculates the required matrices and calibrates the calibrated space so that the simulation box is positioned at the correct place in front of the player. It also updates the simulation pose, scaling the simulation based on the current task. The simulation is not scaled if we are in an Observer Trials task, as this would mess up with the playback.

The SimulationBoxCenter object is a reference for the CalibratedSpaceController as to where the center of the simulation box is (used for calibrating the calibrated space).

Finally, I added the necessary objects in the scene and hooked up the serialized fields.