Show/hide the avatars based on whether the simulation box is visible
The avatars are now nested inside an "avatar parent" game object. This object is enabled and disabled at the same time as the simulation box. For the player, this means that the avatars are only visible if the simulation box is visible.

Previously, the avatars were visible when the player was in the intro and outro to the Trials task (since they are in the shared state at this point). We don't want to destroy the avatars completely, merely hide them for the player.