VR client autoconnects to server and puts simulation in front of player
Edited the game to bypass the normal game logic. Now the game automatically connects to the server when the game starts, and puts the simulation in front of the player. The player can reset this position by clicking the space bar.

Added a README with instructions on how to record.