Clear grab from grabbers when interaction manager is disabled
This ensures that interactions from previous tasks are not being applied at the beginning of the next task. Previously, interactions were being wiped at the end of a task, but then re-applied at the beginning of the next task. If the atom index for this interaction was out of range for the new simulation, then the server threw an error and broke the game.