Added Observer Trials videos
I was having a problem with the videos being black in-game. I think the issue was that the raw image for the VideoPlayerController was not correct (it was serialized, and I think using a different UI's raw image). The Video Player Controller now gets the component from the game object, which seems to have fixed the issue. I don't have access to VR right now though, so I need to check this works when I'm back in the office