Place left button at the center of the left face
Place the left-hand answer now button at the center of the left face. I originally tried it in the same position as the other button (mirrored), but it felt way too hidden, especially as players may be facing more often towards the right of the sim box to see the timer/scores. Placing it at the center made it more visible and easier to push.