Update run_game.py
Cleanup() was being called multiple times. We have now added a check at the beginning to check whether the clean up has already been done, in which case we do not do it again.

This commit also changes the way that we stop the server and puppeteering client. Instead of "terminate", we send a signal (either SIGINT or CTRL_C_EVENT) instead, which is cleaner.