Change number of trials per set to 5
We are going to give the player 5 sets of 5 trials, so we have updated the UI to have 5 icons for the results of each set (it was previously 8 per set).

Also added logic for ensuring that the set number written on the UI is only updated after the first trial of the next set, so that it doesn't show the number of the new set below the answers of the previous set.