Update TrialsTimer.cs to stop first training trial skipping
We have added a few safety catches to ensure that the first trial does not skip straight to answering: make sure the coroutine for the timer animation has finished and explicitly setting the values of the variables. We think this issue might arise from the duration not being set yet. I have tested this out a few times in-game and I haven't run into the bug (but obvs this doesn't mean it is definitely fixed).